+++ package.json
@@ -13,6 +13,7 @@
     "@element-plus/icons-vue": "^2.1.0",
     "axios": "^1.6.0",
     "element-plus": "^2.4.2",
+    "pinia": "2.0.3",
     "vue": "^3.3.4",
     "vue-router": "^4.2.5"
   },
