+++ src/components/camera.vue
@@ -1,0 +1,7 @@
+<template>
+  <div></div>
+</template>
+
+<script lang="ts" setup></script>
+
+<style lang="scss" scoped></style>
