+++ src/components/geo.vue
@@ -1,5 +1,5 @@
 <template>
-  <div>
+  <div class="data-source">
     <el-drawer
       @close="() => (drawer = false)"
       v-model="drawer"
@@ -7,17 +7,42 @@
       direction="btt"
       close-on-click-modal
       show-close
+      destroy-on-close
       size="38%"
     >
-      <data-table></data-table>
+      <data-table :mapId="mapId" :type="type"></data-table>
     </el-drawer>
-    <el-button @click="() => (drawer = true)">平面数据</el-button>
+    <el-button
+      v-if="mapId"
+      v-for="btn in geo_button"
+      @click="() => handleBtnClick(btn)"
+      >{{ btn.label }}</el-button
+    >
   </div>
 </template>
 
 <script lang="ts" setup>
 import dataTable from "./dataTable.vue";
+import { geo_button } from "@/config/geo-button";
 const drawer = ref(false);
+const mapId = ref("");
+const type = ref();
+const route = useRoute();
+
+const handleBtnClick = (btn) => {
+  type.value = btn;
+  drawer.value = true;
+};
+
+watch(
+  () => route.query.id,
+  (id) => {
+    mapId.value = id as string;
+  },
+  {
+    immediate: true,
+  }
+);
 </script>
 
 <style lang="scss" scoped></style>
