+++ src/config/geo-button/index.ts
@@ -1,0 +1,14 @@
+export const geo_button = [
+    {
+        label: '平面数据',
+        value: 'polygon',
+    },
+    {
+        label: '线数据',
+        value: 'polyline',
+    },
+    {
+        label: '点数据',
+        value: 'marker',
+    }
+]
